requirements.txt
# ComfyUI Viewer Dependencies

flask>=2.0.0
flask-socketio>=5.3.0
python-socketio>=5.0.0
requests>=2.25.0
pyyaml>=6.0
websocket-client>=1.0.0
watchdog>=3.0.0
pillow>=10.0.0  # For extracting PNG metadata
